Reorganized package hierarchy
