--- source_files/Danio_test.txt
@@ -1,4 +1,0 @@
-Gene ID	"Gene name"	Anatomical entity ID	"Anatomical entity name"	Developmental stage ID	"Developmental stage name"	Expression	Call quality
-ENSDARG00000000001	"slc35a5"	UBERON:0000468	"multi-cellular organism"	UBERON:0000113	"post-juvenile adult stage"	present	poor quality
-ENSDARG00000000001	"slc35a5"	UBERON:0001997	"olfactory epithelium"	UBERON:0000104	"life cycle"	absent	poor quality
-ENSDARG00000000001	"slc35a5"	ZFA:0001093	"unspecified (Danio)"	UBERON:0000106	"zygote stage"	present	high quality
