--- src/main/java/BgeeExpressionToOWLTestTest.java
@@ -1,14 +1,0 @@
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-public class BgeeExpressionToOWLTestTest {
-
-	@Test
-	public void test() {
-		fail("Not yet implemented");
-
-	}
-
-}
